requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.0
celluloid==0.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588599/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
dominate==2.7.0
executing==1.2.0
fastjsonschema==2.16.2
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.39.2
fqdn==1.5.1
future @ file:///croot/future_1677599870788/work
gh==0.0.4
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.1
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
grpcio==1.51.1
idna @ file:///croot/idna_1666125576474/work
ipykernel==6.21.2
ipython==8.10.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-events==0.5.0
jupyter-ydoc==0.2.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_fileid==0.6.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
kaggle==1.5.12
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib @ file:///croot/matplotlib-suite_1678723872360/work
matplotlib-inline==0.1.6
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
munkres==1.1.4
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nflows==0.14
notebook==6.5.2
notebook_shim==0.2.2
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
opencv-python==4.7.0.68
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.4.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.22.0
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PyQt5-sip==12.11.0
pyrsistent==0.19.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-json-logger==2.0.5
python-slugify==8.0.0
pytorch-fid==0.3.0
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scipy==1.9.3
Send2Trash==1.8.0
sentry-sdk==1.17.0
setproctitle==1.3.2
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
torch==1.13.1
torch-fidelity==0.3.0
torchaudio==0.11.0
torchfile==0.1.0
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1673575502006/work
visdom==0.2.4
wandb==0.14.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
y-py==0.5.5
ypy-websocket==0.8.2
